refactor: move ActionYaml and WorkflowYaml struct to yaml.go
