+++ internal/parse/action.go
@@ -320,40 +320,3 @@ func (r *ActionRuns) String() string {
 	str += fmt.Sprintf("]")
 	return str
 }
-
-type ActionYaml struct {
-	Name        *string                      `yaml:"name"`
-	Description *string                      `yaml:"description"`
-	Inputs      map[string]*actionInputYaml  `yaml:"inputs"`
-	Outputs     map[string]*actionOutputYaml `yaml:"outputs"`
-	Runs        *actionRunsYaml              `yaml:"runs"`
-}
-
-type actionInputYaml struct {
-	Default     *string `mapstructure:"default"`
-	Description *string `mapstructure:"description"`
-	Required    *string `mapstructure:"required"`
-}
-
-type actionOutputYaml struct {
-	Description *string `mapstructure:"description"`
-}
-
-type actionRunsYaml struct {
-	Using string         `yaml:"using"`
-	Steps []*interface{} `yaml:"steps"`
-}
-
-func (y *ActionYaml) inputs() map[string]*actionInputYaml {
-	if y.Inputs == nil {
-		return map[string]*actionInputYaml{}
-	}
-	return y.Inputs
-}
-
-func (y *ActionYaml) outputs() map[string]*actionOutputYaml {
-	if y.Outputs == nil {
-		return map[string]*actionOutputYaml{}
-	}
-	return y.Outputs
-}
